requirements.txt
astroid==2.5.2
autopep8==1.5.6
bidict==0.21.2
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
idna==2.10
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
netifaces==0.10.6
pycodestyle==2.7.0
pylint==2.7.4
python-engineio==4.0.1
python-socketio==5.1.0
requests==2.25.1
six==1.15.0
toml==0.10.2
urllib3==1.26.4
websocket-client==0.58.0
wrapt==1.12.1
